+++ src/main/java/net/geforcemods/securitycraft/ClientHandler.java
@@ -131,8 +131,14 @@ import net.minecraft.client.renderer.ShaderProgram;
 import net.minecraft.client.renderer.blockentity.LecternRenderer;
 import net.minecraft.client.renderer.entity.NoopRenderer;
 import net.minecraft.client.renderer.item.ItemProperties;
+import net.minecraft.client.renderer.texture.TextureAtlasSprite;
+import net.minecraft.client.resources.model.AtlasSet;
+import net.minecraft.client.resources.model.AtlasSet.StitchResult;
 import net.minecraft.client.resources.model.BakedModel;
+import net.minecraft.client.resources.model.BlockModelRotation;
+import net.minecraft.client.resources.model.ModelBakery;
 import net.minecraft.client.resources.model.ModelResourceLocation;
+import net.minecraft.client.resources.model.UnbakedModel;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.BlockPos.MutableBlockPos;
 import net.minecraft.core.Direction;
@@ -250,6 +256,10 @@ public class ClientHandler {
 	});
 	public static final ResourceLocation LINKING_STATE_PROPERTY = SecurityCraft.resLoc("linking_state");
 	private static ShaderProgram frameFeedShader;
+	private static ModelBakery modelBakery;
+	private static Map<ResourceLocation, UnbakedModel> standaloneModels;
+	private static Map<ResourceLocation, AtlasSet.StitchResult> atlasPreperations;
+	private static ModelBakery.TextureGetter textureGetter;
 
 	private ClientHandler() {}
 
@@ -303,8 +313,10 @@ public class ClientHandler {
 		Block sri = SCContent.SECURE_REDSTONE_INTERFACE.get();
 		ResourceLocation sriName = Utils.getRegistryName(sri);
 		ResourceLocation modelBase = sriName.withPrefix("block/");
-		BakedModel poweredSriSender = modelRegistry.get(ModelResourceLocation.standalone(modelBase.withSuffix("_sender_on")));
-		BakedModel poweredSriReceiver = modelRegistry.get(ModelResourceLocation.standalone(modelBase.withSuffix("_receiver_on")));
+		ResourceLocation poweredSriSenderLocation = modelBase.withSuffix("_sender_on");
+		ResourceLocation poweredSriReceiverLocation = modelBase.withSuffix("_receiver_on");
+		UnbakedModel poweredSriSender = standaloneModels.get(poweredSriSenderLocation);
+		UnbakedModel poweredSriReceiver = standaloneModels.get(poweredSriReceiverLocation);
 
 		for (Block block : disguisableBlocks.get()) {
 			for (BlockState state : block.getStateDefinition().getPossibleStates()) {
@@ -314,9 +326,9 @@ public class ClientHandler {
 
 		for (BlockState state : sri.getStateDefinition().getPossibleStates()) {
 			if (state.getValue(SecureRedstoneInterfaceBlock.SENDER))
-				registerDisguisedModel(modelRegistry, state, oldModel -> new SecureRedstoneInterfaceBakedModel(poweredSriSender, oldModel));
+				registerDisguisedModel(modelRegistry, state, oldModel -> new SecureRedstoneInterfaceBakedModel(bakeDirectionalModel(poweredSriSenderLocation, poweredSriSender, state.getValue(SecureRedstoneInterfaceBlock.FACING)), oldModel));
 			else
-				registerDisguisedModel(modelRegistry, state, oldModel -> new SecureRedstoneInterfaceBakedModel(poweredSriReceiver, oldModel));
+				registerDisguisedModel(modelRegistry, state, oldModel -> new SecureRedstoneInterfaceBakedModel(bakeDirectionalModel(poweredSriReceiverLocation, poweredSriReceiver, state.getValue(SecureRedstoneInterfaceBlock.FACING)), oldModel));
 		}
 
 		for (String mine : mines) {
@@ -324,6 +336,10 @@ public class ClientHandler {
 		}
 
 		registerBlockMineModel(event, SecurityCraft.resLoc("quartz_mine"), SecurityCraft.mcResLoc("nether_quartz_ore"));
+		modelBakery = null;
+		standaloneModels = null;
+		atlasPreperations = null;
+		textureGetter = null;
 	}
 
 	private static void registerDisguisedModel(Map<ModelResourceLocation, BakedModel> modelRegistry, BlockState state, Function<BakedModel, IDynamicBakedModel> modelFunction) {
@@ -332,6 +348,18 @@ public class ClientHandler {
 		ModelResourceLocation mrl = new ModelResourceLocation(rl, stateString);
 
 		modelRegistry.put(mrl, modelFunction.apply(modelRegistry.get(mrl)));
+	}
+
+	private static BakedModel bakeDirectionalModel(ResourceLocation location, UnbakedModel model, Direction facing) {
+		try {
+			ModelResourceLocation debugName = new ModelResourceLocation(location, facing.getName());
+
+			return modelBakery.new ModelBakerImpl(textureGetter(), debugName).bakeUncached(model, modelRotationFromDirection(facing));
+		}
+		catch (Exception exception) {
+			SecurityCraft.LOGGER.warn("Unable to bake standalone model: '{}': {}", location, exception);
+			return null;
+		}
 	}
 
 	private static void registerBlockMineModel(ModelEvent.ModifyBakingResult event, ResourceLocation mineRl, ResourceLocation realBlockRl) {
@@ -821,5 +849,45 @@ public class ClientHandler {
 
 	public static ShaderProgram getFrameFeedShader() {
 		return frameFeedShader;
+	}
+
+	public static void setModelBakery(ModelBakery modelBakery) {
+		ClientHandler.modelBakery = modelBakery;
+	}
+
+	public static void setStandaloneModels(Map<ResourceLocation, UnbakedModel> standaloneModels) {
+		ClientHandler.standaloneModels = standaloneModels;
+	}
+
+	public static void setAtlasPreperations(Map<ResourceLocation, StitchResult> atlasPreperations) {
+		ClientHandler.atlasPreperations = atlasPreperations;
+	}
+
+	private static BlockModelRotation modelRotationFromDirection(Direction direction) {
+		return switch (direction) {
+			case DOWN -> BlockModelRotation.X180_Y0;
+			case UP -> BlockModelRotation.X0_Y0;
+			case NORTH -> BlockModelRotation.X90_Y0;
+			case SOUTH -> BlockModelRotation.X90_Y180;
+			case WEST -> BlockModelRotation.X90_Y270;
+			case EAST -> BlockModelRotation.X90_Y90;
+		};
+	}
+
+	private static ModelBakery.TextureGetter textureGetter() {
+		if (textureGetter == null) {
+			textureGetter = (debugName, material) -> {
+				AtlasSet.StitchResult stitchResult = atlasPreperations.get(material.atlasLocation());
+				TextureAtlasSprite sprite = stitchResult.getSprite(material.texture());
+
+				if (sprite != null)
+					return sprite;
+				else
+					return stitchResult.missing();
+			};
+
+		}
+
+		return textureGetter;
 	}
 }