+++ Changelog.md
@@ -3,6 +3,7 @@
 - Fix: Several mod incompatibilities that caused Frame feeds to not show up correctly
 - Fix: Game performance does not improve when breaking frames with an active frame feed
 - Fix: The Level 3 Keycard cannot be crafted
+- Fix: The powered Secure Redstone Interface model looks wrong when placed on the wall or ceiling
 
 --------------------------Changelog for v1.10 of SecurityCraft--------------------------